+++ ricaun.Revit.DB.Shape.Tests/Shape_Arrow_Tests.cs
@@ -1,5 +1,6 @@
 ﻿using Autodesk.Revit.DB;
 using NUnit.Framework;
+using ricaun.Revit.DB.Shape.Extensions;
 using ricaun.Revit.DB.Shape.Tests.Utils;
 
 namespace ricaun.Revit.DB.Shape.Tests
@@ -11,9 +12,33 @@ namespace ricaun.Revit.DB.Shape.Tests
         {
             var solid = ShapeCreator.CreateArrow();
 
-            var volume = 0.0000629;
-            var surface = 0.0151090;
-            AssertUtils.Solid(solid, 14, 8, volume, surface, 1e-6);
+            AssertArrow(solid);
+        }
+
+        [TestCase(1.0)]
+        [TestCase(2.0)]
+        [TestCase(3.0)]
+        [TestCase(4.2)]
+        [TestCase(9.8)]
+        public void CreateArrow_Scale(double scale)
+        {
+            var solid = ShapeCreator.CreateArrow();
+
+            solid = solid.CreateTransformed(Transform.Identity.ScaleBasis(scale));
+
+            AssertArrow(solid, scale);
+        }
+
+
+        private void AssertArrow(Solid solid, double scale = 1.0)
+        {
+            const double Tolerance = 1e-2;
+            const double VolumeArrow = 0.0000629;
+            const double SurfaceArrow = 0.0151090;
+
+            var volume = VolumeArrow * scale * scale * scale;
+            var surface = SurfaceArrow * scale * scale;
+            AssertUtils.Solid(solid, 14, 8, volume, surface, Tolerance);
         }
     }
 }